+++ R/common.R
@@ -27,7 +27,7 @@ INTERNAL <- "internal"
 RESERVED <- c(INDEX, TIME, STEP, STATE, DSTATEDT, STATE_NEXT, USER,
               SPECIAL_LHS, "delay", "dde", INTERNAL)
 RESERVED_PREFIX <- c(SPECIAL_LHS, "odin", "offset", "delay", "interpolate")
-VALID_ARRAY <- c("-", "+", ":", "(", "length", "dim", "[")
+VALID_ARRAY <- c("-", "+", ":", "(", "length", "dim", "[", "as.integer")
 INTERPOLATION_TYPES <- c("constant", "linear", "spline")
 SPECIAL_DATA_TYPES <- c("void", "ring_buffer")
 
@@ -90,10 +90,11 @@ FUNCTIONS <- list(
   lchoose = 2L,
   sign = 1L,
   ## Rounding
-  round = 1L,
+  round = c(1L, 2L),
   trunc = 1L,
   floor = 1L,
   ceil = 1L,
+  as.integer = 1L,
   ## Big pile of trig:
   cos = 1L,   sin = 1L,   tan = 1L,
   acos = 1L,  asin = 1L,  atan = 1L,  atan2 = 2L,
@@ -126,14 +127,19 @@ FUNCTIONS_STOCHASTIC <- list(
   rweibull = 2L, # shape, scale
   rwilcox = 2L, # m, n
   rmultinom = 2L, # n, p
-  rsignrank = 1L # n
+  rsignrank = 1L, # n
+  ## Custom
+  rmhyper = 2L
 )
 
 FUNCTIONS_REWRITE_RF <-
-  grep("_rand$", names(FUNCTIONS_STOCHASTIC), invert = TRUE, value = TRUE)
+  setdiff(
+    grep("_rand$", names(FUNCTIONS_STOCHASTIC), invert = TRUE, value = TRUE),
+    "rmhyper")
 
 FUNCTIONS_INPLACE <- list(
-  rmultinom = list(len = 3L, dest = 4L, type = "int"))
+  rmultinom = list(len = 3L, dest = 4L, type = "int"),
+  rmhyper = list(len = 3L, dest = 4L, type = "int"))
 
 ## Here we need to do a bit of a faff because unary functions need
 ## adding.  This may get tightened up later to either use local() or