+++ R/generate_c_class.R
@@ -47,8 +47,8 @@ odin_c_class <- function(base, core, user, features, dll, ir, package) {
 }
 
 
-odin_c_class_set_user <- function(features, env = .GlobalEnv) {
-  args <- alist("..." =, user = , unused_user_action = NULL)
+odin_c_class_set_user <- function(features, env = emptyenv()) {
+  args <- alist("..." = , user = , unused_user_action = NULL) # nolint
   args[[2]] <- quote(list(...))
 
   check_user <- quote(
@@ -72,9 +72,9 @@ odin_c_class_set_user <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_initial <- function(features, env = .GlobalEnv) {
+odin_c_class_initial <- function(features, env = emptyenv()) {
   time <- if (features$discrete) STEP else TIME
-  args <- alist(time =)
+  args <- alist(time = ) # nolint
   names(args) <- time
   if (features$initial_time_dependent) {
     time_clean <- if (features$discrete) "as_integer" else "as_numeric"
@@ -88,9 +88,9 @@ odin_c_class_initial <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_update <- function(features, env = .GlobalEnv) {
+odin_c_class_update <- function(features, env = emptyenv()) {
   if (features$discrete) {
-    args <- alist(step =, y =)
+    args <- alist(step = , y = ) # nolint
     if (features$has_delay) {
       body <- quote(stop("Can't call update() on delay models"))
     } else {
@@ -105,11 +105,11 @@ odin_c_class_update <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_deriv <- function(features, env = .GlobalEnv) {
+odin_c_class_deriv <- function(features, env = emptyenv()) {
   if (features$discrete) {
     NULL
   } else {
-    args <- alist(t =, y =)
+    args <- alist(t = , y = ) # nolint
     body <- call(".Call", quote(private$core$rhs_r), quote(private$ptr),
                  quote(as_numeric(t)), quote(as_numeric(y)),
                  PACKAGE = quote(private$dll))
@@ -118,7 +118,7 @@ odin_c_class_deriv <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_run <- function(features, env = .GlobalEnv) {
+odin_c_class_run <- function(features, env = emptyenv()) {
   if (features$discrete) {
     odin_c_class_run_discrete(features, env)
   } else {
@@ -127,8 +127,9 @@ odin_c_class_run <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_run_continuous <- function(features, env = .GlobalEnv) {
-  args <- alist(t =, y = NULL, "..." =, use_names = TRUE, tcrit = NULL)
+odin_c_class_run_continuous <- function(features, env = emptyenv()) {
+  args <- alist(t = , y = NULL, "..." = , use_names = TRUE, # nolint
+                tcrit = NULL)
   if (features$has_delay) {
     args <- c(args, alist(n_history = DEFAULT_HISTORY_SIZE))
   }
@@ -182,8 +183,9 @@ odin_c_class_run_continuous <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_run_discrete <- function(features, env = .GlobalEnv) {
-  args <- alist(step =, y = NULL, "..." =, use_names = TRUE, replicate = NULL)
+odin_c_class_run_discrete <- function(features, env = emptyenv()) {
+  args <- alist(step = , y = NULL, "..." = , use_names = TRUE, # nolint
+                replicate = NULL)
 
   check_step <- quote(step <- as_integer(step))
   check_y <- r_expr_if(quote(is.null(y)), quote(y <- self$initial(step)))
@@ -214,7 +216,7 @@ odin_c_class_run_discrete <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_update_metadata <- function(features, env = .GlobalEnv) {
+odin_c_class_update_metadata <- function(features, env = emptyenv()) {
   body <- list(
     call("<-", quote(meta),
          call(".Call", quote(private$core$metadata), quote(private$ptr),
@@ -230,7 +232,7 @@ odin_c_class_update_metadata <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_initialize <- function(features, env = .GlobalEnv) {
+odin_c_class_initialize <- function(features, env = emptyenv()) {
   args <- alist(user = NULL, unused_user_action = NULL)
   if (features$discrete) {
     set_use_dde <- NULL
@@ -252,15 +254,15 @@ odin_c_class_initialize <- function(features, env = .GlobalEnv) {
 }
 
 
-odin_c_class_contents <- function(features, env = .GlobalEnv) {
+odin_c_class_contents <- function(features, env = emptyenv()) {
   body <- call(".Call", quote(private$core$contents), quote(private$ptr),
                PACKAGE = quote(private$dll))
   as_function(alist(), r_expr_block(body), env)
 }
 
 
-odin_c_class_transform <- function(features, env = .GlobalEnv) {
-  args <- alist(y =)
+odin_c_class_transform <- function(features, env = emptyenv()) {
+  args <- alist(y = ) # nolint
   body <- call("support_transform_variables", quote(y), quote(private))
   as_function(args, r_expr_block(body), env)
 }