+++ inst/library.c
@@ -47,7 +47,7 @@ int user_get_scalar_int(SEXP user, const char *name,
 SEXP user_list_element(SEXP list, const char *name) {
   SEXP ret = R_NilValue, names = getAttrib(list, R_NamesSymbol);
   for (int i = 0; i < length(list); ++i) {
-    if(strcmp(CHAR(STRING_ELT(names, i)), name) == 0) {
+    if (strcmp(CHAR(STRING_ELT(names, i)), name) == 0) {
       ret = VECTOR_ELT(list, i);
       break;
     }
@@ -302,6 +302,13 @@ double odin_sum1(double *x, size_t from, size_t to) {
   return tot;
 }
 
+int odin_isum1(int *x, size_t from, size_t to) {
+  int tot = 0.0;
+  for (size_t i = from; i < to; ++i) {
+    tot += x[i];
+  }
+  return tot;
+}
 
 void lagvalue_ds(double t, int *idx, int dim_idx, double *state) {
   typedef void (*lagvalue_type)(double, int*, int, double*);
@@ -345,4 +352,38 @@ void interpolate_check_y(size_t nx, size_t ny, size_t i, const char *name_arg, c
                i, name_arg, nx, name_target);
     }
   }
+}
+
+
+void rmhyper(size_t n_sample, int *k, size_t m) {
+  int N = 0;
+  for (size_t i = 0; i < m; ++i) {
+    N += k[i];
+  }
+  if ((int)n_sample > N) {
+    Rf_error("Requesting too many elements in rmhyper (%d from %d)",
+             n_sample, N);
+  }
+  int k_other = N - k[0];
+  k[0] = Rf_rhyper(k[0], k_other, n_sample);
+  for (size_t i = 1; i < m - 1; ++i) {
+    k_other -= k[i];
+    n_sample -= k[i - 1];
+    k[i] = Rf_rhyper(k[i], k_other, n_sample);
+  }
+  k[m - 1] = n_sample - k[m - 2];
+}
+
+
+void rmhyper_i(size_t n_sample, int *k, size_t m, int *ret) {
+  memcpy(ret, k, m * sizeof(int));
+  rmhyper(n_sample, ret, m);
+}
+
+
+void rmhyper_d(size_t n_sample, double *k, size_t m, int *ret) {
+  for (size_t i = 0; i < m; ++i) {
+    ret[i] = k[i];
+  }
+  rmhyper(n_sample, ret, m);
 }